pyproject.toml
[project]
name = "vel-harness"
version = "0.1.0"
description = "Deep agent harness for data, decisions, and research"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Product" }
]
keywords = ["agent", "llm", "ai", "automation", "data-analysis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "vel-ai @ git+https://github.com/rscheiwe/vel.git@main",
    "agentmesh-py @ git+https://github.com/rscheiwe/mesh.git@main",
    "pyyaml>=6.0",
    "textual>=0.47.0",
    "click>=8.1.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
database = ["asyncpg>=0.29.0"]
accurate-tokens = ["tiktoken>=0.5.0"]
ctxzip = ["ctxzippy>=0.1.0"]
all = ["vel-harness[database,accurate-tokens,ctxzip]"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "black>=24.0",
    "ruff>=0.1.0",
    "mypy>=1.8",
]

[project.scripts]
vel-harness = "vel_harness.cli.main:cli_entry"
valis = "valis_cli.main:main"

[project.urls]
Homepage = "https://github.com/example/vel-harness"
Documentation = "https://github.com/example/vel-harness#readme"
Repository = "https://github.com/example/vel-harness"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["vel_harness", "valis_cli"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
